+++ lua/astrocommunity/editing-support/other-nvim/init.lua
@@ -1,0 +1,17 @@
+return {
+  "rgroli/other.nvim",
+  ft = {
+    -- built-in mapping support
+    "livewire",
+    "angular",
+    "laravel",
+    "rails",
+    "golang",
+    "python",
+    "react",
+    "rust",
+    "elixir",
+    "clojure",
+  },
+  opts = {},
+}