+++ lua/astrocommunity/editing-support/rainbow-delimiters-nvim/init.lua
@@ -1,6 +1,7 @@
 return {
   {
     "HiPhish/rainbow-delimiters.nvim",
+    submodules = false,
     dependencies = {
       "nvim-treesitter/nvim-treesitter",
       {