+++ lua/astrocommunity/note-taking/neorg/init.lua
@@ -1,6 +1,6 @@
 return {
   "nvim-neorg/neorg",
-  version = "^8",
+  version = "^9",
   event = "VeryLazy",
   dependencies = { "nvim-treesitter/nvim-treesitter" },
   opts = function(_, opts)
@@ -29,4 +29,15 @@ return {
       },
     })
   end,
+  specs = {
+    {
+      "nvim-treesitter/nvim-treesitter",
+      opts = function(_, opts)
+        if opts.ensure_installed ~= "all" then
+          opts.ensure_installed =
+            require("astrocore").list_insert_unique(opts.ensure_installed, { "norg", "norg_meta" })
+        end
+      end,
+    },
+  },
 }